patch: updated deployments
